+++ changelog.md
@@ -1,3 +1,8 @@
+# 2.0.0
+- Added auto code Repo (My life and geode mods life is so much easier now)
+- Added a tutorial because many people don't know how to use the mod lol
+- Added a secret codes mode so you don't know the codes 
+- Added soggy (i recommend soggy code mod for best experience)
 # 1.2.1
 - Added newest code.
 - Made everything better.